conflate shared service middleware
