+++ swift_browser_ui/common/__init__.py
@@ -1,0 +1,1 @@
+"""Functionality common across the swift_browser_ui package services."""
