+++ swift_browser_ui/common/types.py
@@ -1,0 +1,12 @@
+"""Common types in swift_browser_ui project."""
+
+
+import typing
+
+import aiohttp.web
+
+
+AiohttpHandler = typing.Callable[
+    [aiohttp.web.Request],
+    typing.Coroutine[typing.Awaitable, typing.Any, aiohttp.web.Response],
+]
